dashboard/requirements.txt
dash==2.10.2
plotly==5.14.1
psycopg2-binary==2.9.6
sqlalchemy==2.0.15
python-dotenv==1.0.0
numpy<2.0
pandas>=1.5.3,<2.2.2
